phase1.py
from scipy.io import loadmat 
import matplotlib.pyplot as plt 
import matplotlib 
import matplotlib as mpl 
from skimage import feature 
from skimage import filters 
from skimage.color.adapt_rgb import adapt_rgb, each_channel, hsv_value 
from skimage.color import rgb2gray 
from sklearn.metrics import pairwise_distances 
from sklearn.preprocessing import StandardScaler 
from sklearn.feature_selection import mutual_info_classif, chi2 
import random 
import numpy as np
import cv2
from PIL import Image